requirements.txt
requests>=2.28.0
ttkbootstrap>=1.10.0
yt-dlp>=2023.3.4
mutagen>=1.46.0
plyer>=2.1.0
spotipy>=2.23.0
watchdog>=2.3.1
pillow>=9.5.0
matplotlib>=3.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
